bug fix (invalid narrowing conversion from "int" to "float")

+++ main.cpp
@@ -7,7 +7,7 @@ using namespace std;
 Color green = {173, 204, 96, 255};
 Color darkGreen = {43, 51, 24, 255};
 
-int cellSize = 30;
+float cellSize = 30;
 int cellCount = 25;
 
 class Snake
